+++ app/src/main/java/com/concordium/wallet/App.kt
@@ -4,8 +4,6 @@ import android.app.Application
 import android.content.Context
 import com.concordium.wallet.data.backend.ws.WsCreds
 import com.concordium.wallet.util.Log
-import com.google.firebase.analytics.FirebaseAnalytics
-import com.google.firebase.crashlytics.FirebaseCrashlytics
 import com.walletconnect.android.Core
 import com.walletconnect.android.CoreClient
 import com.walletconnect.android.relay.ConnectionType
@@ -15,7 +13,6 @@ import com.walletconnect.sign.client.SignClient
 import java.util.concurrent.TimeUnit
 
 class App : Application() {
-
     companion object {
         lateinit var appContext: Context
         lateinit var appCore: AppCore
@@ -36,7 +33,7 @@ class App : Application() {
         appContext = this
         initAppCore()
         initWalletConnect()
-        initFirebase()
+        appCore.tracker.installation(this)
     }
 
     fun initAppCore() {
@@ -74,18 +71,5 @@ class App : Application() {
                 println("WalletConnect -> SIGN ERROR ${error.throwable.stackTraceToString()}")
             }
         )
-    }
-
-    private fun initFirebase() {
-        with(FirebaseCrashlytics.getInstance()) {
-            setCrashlyticsCollectionEnabled(!BuildConfig.DEBUG)
-            if (BuildConfig.DEBUG) {
-                deleteUnsentReports()
-            }
-        }
-
-        with(FirebaseAnalytics.getInstance(this)) {
-            setAnalyticsCollectionEnabled(!BuildConfig.DEBUG)
-        }
     }
 }