+++ app/src/main/java/com/concordium/wallet/AppCore.kt
@@ -7,6 +7,9 @@ import com.concordium.wallet.core.crypto.CryptoLibrary
 import com.concordium.wallet.core.crypto.CryptoLibraryReal
 import com.concordium.wallet.core.gson.BigIntegerTypeAdapter
 import com.concordium.wallet.core.gson.RawJsonTypeAdapter
+import com.concordium.wallet.core.tracking.AppTracker
+import com.concordium.wallet.core.tracking.MatomoAppTracker
+import com.concordium.wallet.core.tracking.NoOpAppTracker
 import com.concordium.wallet.data.backend.ProxyBackend
 import com.concordium.wallet.data.backend.ProxyBackendConfig
 import com.concordium.wallet.data.backend.airdrop.AirDropBackend
@@ -14,9 +17,12 @@ import com.concordium.wallet.data.backend.airdrop.AirDropBackendConfig
 import com.concordium.wallet.data.backend.tokens.TokensBackend
 import com.concordium.wallet.data.backend.tokens.TokensBackendConfig
 import com.concordium.wallet.data.model.RawJson
+import com.concordium.wallet.data.preferences.TrackingPreferences
 import com.concordium.wallet.data.room.Identity
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
+import org.matomo.sdk.Matomo
+import org.matomo.sdk.TrackerBuilder
 import java.math.BigInteger
 
 class AppCore(val context: Context) {
@@ -26,10 +32,26 @@ class AppCore(val context: Context) {
     private val tokenBackendConfig = TokensBackendConfig(gson)
     private val airdropBackendConfig = AirDropBackendConfig(gson)
     val cryptoLibrary: CryptoLibrary = CryptoLibraryReal(gson)
+
+    private val trackingPreferences = TrackingPreferences(context)
+    private val noOpAppTracker: AppTracker = NoOpAppTracker()
+    private val matomoAppTracker: AppTracker by lazy {
+        TrackerBuilder.createDefault("https://concordium.matomo.cloud/matomo.php", 8)
+            .build(Matomo.getInstance(context))
+            .let(::MatomoAppTracker)
+    }
+    val tracker: AppTracker
+        get() =
+            if (trackingPreferences.isTrackingEnabled)
+                matomoAppTracker
+            else
+                noOpAppTracker
+
     val session: Session = Session(App.appContext)
     var newIdentities = mutableMapOf<Int, Identity>()
 
-    private val authenticationManagerGeneric: AuthenticationManager = AuthenticationManager(session.getBiometricAuthKeyName())
+    private val authenticationManagerGeneric: AuthenticationManager =
+        AuthenticationManager(session.getBiometricAuthKeyName())
     private var authenticationManagerReset: AuthenticationManager = authenticationManagerGeneric
     private var authenticationManager: AuthenticationManager = authenticationManagerGeneric
     private var resetBiometricKeyNameAppendix: String = ""