+++ app/src/main/java/com/concordium/wallet/ui/passphrase/setup/OneStepSetupWalletViewModel.kt
@@ -27,6 +27,7 @@ class OneStepSetupWalletViewModel(application: Application) : AndroidViewModel(a
     val eventsFlow: Flow<Event> = mutableEventsFlow
 
     init {
+        App.appCore.tracker.welcomePhrase()
         generatePhrase()
     }