+++ app/src/main/java/com/concordium/wallet/ui/welcome/WelcomeNotificationPermissionDialog.kt
@@ -10,15 +10,22 @@ import androidx.activity.result.contract.ActivityResultContracts
 import androidx.activity.result.registerForActivityResult
 import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatDialogFragment
+import androidx.lifecycle.lifecycleScope
 import com.concordium.wallet.R
+import com.concordium.wallet.data.preferences.NotificationPreferences
 import com.concordium.wallet.databinding.DialogWelcomeNotificationPermissionBinding
 import com.concordium.wallet.util.Log
+import kotlinx.coroutines.delay
 
 class WelcomeNotificationPermissionDialog : AppCompatDialogFragment() {
     override fun getTheme(): Int =
         R.style.CCX_Dialog
 
     private lateinit var binding: DialogWelcomeNotificationPermissionBinding
+
+    private val notificationPreferences: NotificationPreferences by lazy {
+        NotificationPreferences(requireContext())
+    }
 
     @RequiresApi(Build.VERSION_CODES.TIRAMISU)
     private val notificationPermissionsLauncher =
@@ -59,6 +66,12 @@ class WelcomeNotificationPermissionDialog : AppCompatDialogFragment() {
                 onNotificationPermissionResult(isGranted = true)
             }
         }
+
+        // Track showing the dialog once it is visible to the user.
+        viewLifecycleOwner.lifecycleScope.launchWhenStarted {
+            delay(500)
+            notificationPreferences.hasEverShownPermissionDialog = true
+        }
     }
 
     private fun onNotificationPermissionResult(isGranted: Boolean) {
@@ -68,8 +81,6 @@ class WelcomeNotificationPermissionDialog : AppCompatDialogFragment() {
         )
 
         dismiss()
-        (requireActivity() as? WelcomeNotificationPermissionDialogListener)
-            ?.onNotificationPermissionDialogDismissed(isGranted)
     }
 
     companion object {