+++ app/src/main/java/com/concordium/wallet/ui/welcome/WelcomePromoViewModel.kt
@@ -5,6 +5,7 @@ import android.os.Build
 import androidx.lifecycle.AndroidViewModel
 import com.concordium.wallet.App
 import com.concordium.wallet.data.preferences.NotificationPreferences
+import com.concordium.wallet.data.preferences.TrackingPreferences
 import java.math.BigInteger
 
 class WelcomePromoViewModel(application: Application) : AndroidViewModel(application) {
@@ -12,11 +13,16 @@ class WelcomePromoViewModel(application: Application) : AndroidViewModel(applica
     private val notificationPreferences: NotificationPreferences by lazy {
         NotificationPreferences(application)
     }
+    private val trackingPreferences: TrackingPreferences by lazy {
+        TrackingPreferences(application)
+    }
 
     val shouldShowNotificationPermissionDialog: Boolean
         get() = IS_SENDING_ANY_NOTIFICATIONS
                 && Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU
                 && !notificationPreferences.hasEverShownPermissionDialog
+    val shouldShowTrackingPermissionDialog: Boolean
+        get() = !trackingPreferences.hasDecidedOnPermission
     val shouldSetUpPassword: Boolean
         get() = !App.appCore.session.hasSetupPassword
     val accountActivationReward = BigInteger("1000000000")
@@ -25,10 +31,6 @@ class WelcomePromoViewModel(application: Application) : AndroidViewModel(applica
 
     fun initialize() {
         App.appCore.session.startedInitialSetup()
-    }
-
-    fun onNotificationPermissionDialogDismissed() {
-        notificationPreferences.hasEverShownPermissionDialog = true
     }
 
     private companion object {